added basic fr_fr and zh_cn lang files
rudimentary language translations for french and chinese. these should get by for now, but bound to be at least a dozen grammatical errors in the blocks section as i'm not at all a proficient native speaker. pokemon item translation errors should be minimal 
thank you to yacinouvn and cjbreizh; xionggapuo and justpost, respectively, for helping assist with numerous translation questions!

not sure if the tm and .tooltip translations can be properly validated in-game at the moment seemingly due to bugs, .desc ones look to be working